Large initial commit of the mu2e tracker production software. This constitutes most of the code that was in the Production_Environment folder as of 20200728-ish.
There are three primary directories:
1. Database -- code that interfaces between the data collection GUI
and the database.
2. Modules -- utility classes, most notably the automerger, which
periodically merges the local database with that on the network.
3. GUIS -- the panel production GUI and database-viewer GUI. Within
the panel GUI there is also the code for the external arduino-based
GUIS.

As a standalone package, a yet-to-be-committed setup script will need
to set some environment variables and copy requisite source images
and data to the sandbox.