PANGUI uses absolute pathnames, determined from current directory. Instead of relative path locations use absolute when determining the location of modules, data, and files. And instead of getting these from hardcoded variables, determine dynamically.
This job isn't complete -- still need to do this for database
(local and network) locations and for txtprocessor, which
duplicates the code that I'm here changing in PANGUI. No reason
it should be done twice.

The former, at least, must be completed before this code is
ready for launch to the lab and simultaneous development.

Abandoning the environment variable solution in setup.bat. In a
NIX system it's easy to set and use env variables. Not so in
Windows.