Mergedown now only syncs data dir and database.db.
Now that we're going to be running from git, the mergedown
as-is would copy a bunch of junk from the network PE into
the lab sandbox that we don't want to commit to the repo.

So we have two options: (1) keep mergedown the same but
remove all the junk from the network or (2) change the
mergedown to only copy the stuff that won't ever be saved
to the repo anyway (viz. database.db and data/).

I opted for (2), not ready yet, to mess with network PE
area as long as I'm still migrating computers to the git
system. Ultimately we'll want to clean out the network PE
so that it conly contains data.