Add panel temp measurement backend.
Change some function names. Change savePanelTemp function signature
to only record two temperatures. Which paas corresponds to which
temperature can be inferred from procedure id. And temp1 is always
paas-A.

The databaseClasses record function is not copied within each of
processes 1, 2, and 6. Instead it's being defined within the parent
class, PanelProcess, or whatever it's called.

Adding the table into the DB is very straightforward, and I'm
going to test the full stack presently.