Tested panel heating with fake data.
Fixed typos, save csv data to Data/ directory, actually show the window
when it launches.

Tested the heating GUI from within the PANGUI. Correctly saves to the DB
and the CSV files. More tests with real data, processes 2 and 6, and
the PAAS-C, and NONE option.

The arduino code for this to work is currently only installed on one
heating box. Kate is still tweaking the ino and GUI. Doesn't correctly
work with the beginning of process 2 at the moment.