Move lpal (#11)
* Moved LPAL stuff from pro 2 to 1.
I'm not sure if that's exactly what we want it to look like, but I plan on making things look much more even and clean with the gui_layout branch.

* Removed check straws button, since it doesn't do anything.

* Made PANGUI changes
Changed data count for pro 1 & 2
Moved pallet code widgets from pro 2 to pro 1, this shows up in the update, parse, and reset functions for both pro 1 & 2
Removed pallet code widgets from pro2part1
Changed inputs to be verified in pro2part1
Subtracted 2 from all data indicies above 2, since I removed 1 & 2

* Backend changes to save LPAL from pro 1 instead of 2

I moved a few functions from:
Pan2Procedure to Pan1Procedure in databaseClasses
saveDataProcess2 to saveDataProcess1 in dataProcessor.
I have made necessary changes in data lists. Everything looks good now, I will test this in a lab computer!

* LPAL saving functionality added to day 1

While running tests, this branch crashed. LPALs are saved in straw_location table but turns out the lpal data is also saved in database on day 2 as foreign link. This calls for small changes in the DB.

* LPAL doesn't need to be entered to start Pro 1, and validate straws button added.
In order for the finish button to be enabled on pro 1, all the steps must be checked, and the straws must be validated.  The user can check all the steps and validate straws in any order (either can be first), but both must be done.
For this, I added a button to pro 1 and a funciton - checkLPALs().

* Bugfix: before the first updatedata() is called, self.data is a list of lists.  When self.data was referenced in pro1part1(), it was trying to get the 23 index from the first list, thinking that there was only one list.

* Debugged straw validator/steps list bugs.  Some indicies were out of range because I kept confusing the list of input widgets with self.data.

* More debugging, see previous commit.

* UI fix:
-the LPAL entry widgets were HUGE
-now they're capped to 50 pixels

* Forgot to use pyuic5 on panel.ui for previous commit

Co-authored-by: Himanshu Joshi <himanshujoshi790@gmail.com>