Get straw tensioner gui working in standalone mode.
The saveMethod function will not be defined in standalone
mode so don't assume its existence. IF saveMethod is defined
then use it.

The todo would probably be to warn the user if saveMethod
isn't defined AND the gui was launched from within PANGUI.
Maybe a check like if __name__ == '__main__' would do.