Network Storage Migration (#25)
We are exchanging the network computer file storage \\spa-mu2e-network\Files\Data with a simple, dedicated file storage area \\rds01.storage.umn.edu\cse_spa_mu2e\Data.

The network computer ran no code, and only held data, so replacing it with simple file storage is cool.

As far as backups go, instead of doing a nightly backup to \\files.umn.edu\spa\groups\mu2e, we'll rely on rds's "snapshots" which are taken thrice daily. They start falling off after a month, so we'll grab a snapshot on a monthly basis and store it somewhere that won't itself be duplicated in snapshots (maybe an external HD).

The only code changes to this branch are in the setup script (which direct PANGUI to the correct merge-to location), mergedown, and DBViewer. Simple changing of filepaths.

* Change setup to use new rds01 network storage.

Hope to replace spa-mu2e-network --> //files.umn.edu/spa/groups/mu2e/
with a single network storage location //rds01.storage.umn.edu/cse_spa_mu2e
and cut out a network machine altogether. Our embedded database
system should work well with this system.

With this commit, I just point the network location to the
new area. So far so good. Next is to change mergedown and
DBViewer if they aren't automatically changed.

* Point mergedown and dbviewer to new storage location.

Reminder that DBViewer needs to mount the network
drive in order to open in read-only mode. Assume it's
mounted to X:.