Fix pro2 data list indices.
parsePro2 data looks for certain data in certain places in its
data list. When we removed the lpal data from this list, we need
to shift the indices around. This shift got missed for the paas
B temperature data. Fixed. I think this is important because
dataProcessor needs to look in this list for the particular
data.

I also updated the data count for each process in _init_data_lists
which fills the data containers with nulls. The number of data
points is wrong for some of the processes. Evidently not mattering
too much. But if things aren't working should remember this.

I only assumed that this fix will solve problems. Attempting to
test pro 1 and pro 2 with a heater box now.