Updated straw Verify scripts to work as intended. (#38)
* Updated Verify scripts to work as intended.

addToStorage.py was not in the repository. This scripts is needed to add
new straws to storage.csv, which is the file that verifyStorageStraw.py reads from.

Modified the dictionary used in verifyStorageStraw.py to match the naming for the resistance's csv
io, ii, oi, and oo instead if input-output, input-input, etc

Added nomimal_lengths.csv. verifyStorageStraw.py uses the numbers from this file to ensure
the straws are the right lenght and the file was missing.

Updated other broken path.

* Added an import to addToStorage and modified the path for the csv file in verifyStorageStraw

Co-authored-by: Ben Messerly <bamesserly@gmail.com>