New move lpal (#31)
Move recording of LPAL from beginning of process 2 to end of process 1.
--------------------------------------------------------
* Squashed changes from fraught move_lpal onto this new branch.

* Tested LPAL
- I had to use pyuic5 on the ui file to get panel.py
- After that all was well!

* Fix pro2 data list indices.

parsePro2 data looks for certain data in certain places in its
data list. When we removed the lpal data from this list, we need
to shift the indices around. This shift got missed for the paas
B temperature data. Fixed. I think this is important because
dataProcessor needs to look in this list for the particular
data.

I also updated the data count for each process in _init_data_lists
which fills the data containers with nulls. The number of data
points is wrong for some of the processes. Evidently not mattering
too much. But if things aren't working should remember this.

I only assumed that this fix will solve problems. Attempting to
test pro 1 and pro 2 with a heater box now.

* More fixing of data list indices.

TextDataProcessor had a crash related to the number of
data points it was expecting for pro1 and/or pro2.

Similar issue for PANGUI.

Currently, process 2 seems to be broken in that I cannot
unlock the epoxy batches and timers. Hoping @adamarnett
can save the day.

* Fixed Locked Fields Bugs
- Changed the validated indicies on some of the functions for pro 2
- Said functions were checking the wrong widgets

* Section Header Format Changes
- Section headers got tabbed over to match the classes indentation, which broke the section collapsing functionality
- Undid that so it works again

* Reverting PANGUI to 3a8ab72.

* Re-fixing PANGUI locked fields bugs of 5eb4c48.

* Run the formatter on panel.py.

Co-authored-by: Ben Messerly <bamesserly@gmail.com>