Fix to be able to leak test straws a second time.
Straws may get leak tested, be subsequently removed from
production for some reason, and then re-cut or otherwise
re-introduced into production under a new CPAL and CPALID.

Previously, the code would look for record of the straw's
previous tests in the first CPAL it found. Now that straws
may be associated with multiple CPALs, we must check all
associated CPALs, and then we want to return the most recent
CPAL (i.e. the one with the largest value).

So that's what we do.

Another small fix: when looping through test records for old
CPALs, empty lines are introduced when a straw is removed
(I think). The correct way to check for an empty line is
apparently not line == ''. Instead I check that its empty
after stripping away whitespace.

Finally, add a leading backslash to some filepaths which
allows me to launch the GUI from the command line.