Hv standalone(#58)
The HV current measurement is now a popup window, with a launch button for processes 3 and 6.

You can select left or right side, voltage setting.

* Big commit.  Moved changes from hv_standalone to new branch.

* Saves/Loads from DB with a catch.
Procedure 5 is now RIP.  I kept getting an error relating to multiple
instances or references to a single table, since the hv meassurement
class was in the pan5 procedure class and I needed it in 3 and 6 too.
My fix was to move the measurement class out of pan 5 then comment
out the pan 5 class.  That worked for 3 and 6 but totally broke pro 5.
I'll still try to fix it, but 3 and 6 work for now, so that's good.

* Saves and loads!  No catch!
Moving the MeasurementPan5 class out of Pan5Procedure class fixed my
previous commit's problem.  Now Pan5Procedure deals with HV
measurements the same way as 3 and 6.  I also removed some filter
in the function that parses HV data that seemed unnecessary to me.

* Debugging and fine tuning of backwards compatability.
Pro 5 can still edit/save panels, it just doesn't include voltage.
A new feature it has is that you automatically jump to the next straw
on the side you're using with tab or enter.  This prevents it from
saving twice per line since the editingFinished signal triggers the
program to save.

* Added window re-launch ability

* CSV filename generation typos fixed. Added logger.

Co-authored-by: Ben Messerly <bamesserly@gmail.com>