Straw Leak GUI Find Latest CPAL (#61)
When searching and saving data for a straw, use its latest CPAL.

Remove old upload-to-mu2e-cart function.

* Fix to be able to leak test straws a second time.

Straws may get leak tested, be subsequently removed from
production for some reason, and then re-cut or otherwise
re-introduced into production under a new CPAL and CPALID.

Previously, the code would look for record of the straw's
previous tests in the first CPAL it found. Now that straws
may be associated with multiple CPALs, we must check all
associated CPALs, and then we want to return the most recent
CPAL (i.e. the one with the largest value).

So that's what we do.

Another small fix: when looping through test records for old
CPALs, empty lines are introduced when a straw is removed
(I think). The correct way to check for an empty line is
apparently not line == ''. Instead I check that its empty
after stripping away whitespace.

Finally, add a leading backslash to some filepaths which
allows me to launch the GUI from the command line.

* Don't call deprecated strawUpload function during unload.

Also: remove print statements from FindCPAL. For better or for worse,
FindCPAL gets called continually in a thread, so this print mesg was
spammed.

* Fix bat file python command.

I don't understand when or why the .\ is needed.

* FindCPAL should look for the largest CPAL not the largest CPALID.

Also, I checked that python correctly compares strings the way
we want with the max() function.