Promising attempt at lock fix.
Adam's debug messages revealed that automerges take 8-12 seconds!
AFAICT the network DB is indeed locked during this time.

Automerge makes a DB connect()-ion, which has a default timeout of
5 seconds. When I ramped this down to 1 or 2 sec and opened two
instances one after the other I got a locked error! When I ramped
the timeout up to 15 seconds, it went away!

15 seconds should be OK b/c automerges only run every 10 minutes, so
we won't have multiple AM's queuing up on the same computer.

Still possible to have a 3-computer-AM logjam, but that's going to
be very unlikely.

Feeling good that this will solve 99 out of 100 locked problems.