FNAL Mu2e GPVM compatibility (#101)
Changes to DBV and to repo installation (setup.py and a new requirements.txt) are all we need to get the DBV working on gpvms.

This could be tested more thoroughly, but we'll do it live, because FNAL folks want this ASAP, and it doesn't actually affect tracker production.

There is now a dbvDatabaseLocation.txt that the DBV shalt read from to find the db to access. Setup script now takes a couple option flags.

* Clean up setup.py. Introduce dbvDatabasePath.txt.

requirements.txt modified for specifically running on gpvms. Will need
to merge requirements with master.

* Just make a separate requirements file for the gpvms.

* Remove erroneous erase message that cut setup short.

Catch pre-commit setup. Don't really know what Popen might raise.

Add dbv location txt to ignore.

* We need a data/__init__.py for the DBV apparently.