Pro 8 requests (#107)
* Keep parts enabled + bugfix
As long as there's a running process, parts CAN be changed.  They will
automatically save whenever the user finishes editing one, whenever
the submit covers/rings buttons have been pressed, or whenever the save
data funciton is called.  They are also set as enabled whenever absolutely
ANYTHING is clicked in pro 8.  The only time they're disabled is when
a process hasn't been started yet, or a process was paused/finished.

Seperately, I fixed a bug where the gui would crash if something that
wasn't an int was entered into the position field for the bad wire/straw
form and the submit button was clicked.  This was fixed by adding a regex
to the line edit.

* Added previous failures box
- now displays previous bad wire/straw failures
- required a few methods be added to the database processor/classes
- also minor ui tweaks (color of seperator line in pro 8)