Panel resist arduino search (#116)
Problem: computer 15 couldn't find the resistance arduino box device.

The way the device search goes is that it looks for a device with the keyword "USB" in it.

Computer 15 always has the temp/humidity device plugged into it, and this device was being found before the resist arduino.

To fix this, I required that the keyword "arduino" also be found in the device description.

Well this fix worked for computer 15/that specific resistance arduino, BUT it broke all the other computers/resistance arduinos. I'm not sure if the other arduinos are described differently inherently (without the keyword "arduino") OR whether the computer names them differently or what.

So the current solution is to treat computer 15 differently -- having it look for "arduino", but not having the other computers require that keyword.

The real solution is probably to search among the device ID numbers for the resist boxes and use those. I don't have all of them in front of me, so this is the temporary solution.

* resistance arduino goes by different names.

Apparently the port doesn't always have the word 'arduino' in it.

I'm going to need to come up with a more sophisticated solution --
for arbitrary box-by-box, computer-by-computer port descriptions.

In the meantime, a hotfix branch to keep production working.

* Treat the QC computer differently.