Break out individual procedure classes.
Starting a PanXProcedure invokes Procedure functions, which loop through
Procedure's subclasses to find the PanXProcedure that called it in the
first place. In order to separate PanXProcedure from Procedure, I
therefore needed to tell Procedure that PanXProcedure existed and is its
child class.

Obvious in hindsight, but due to this weird subclass looping strategy,
the error messages were really hard to interpret.

Just because this was a fun and rewarding myster to solve:
The highest level error was that PanelProcedure had been deleted or was
not present. Those were red herrings. The real error, lower down, was
that PanelProcedure had incompatible polymorphic identity with
PanXProcedure, which is clear enough, except that it didn't say
PanXProcedure, but instead used it's station id: panX.

tl;dr: needed to import PanXProcedure subclasses into the Procedure
file.