Start saving to DB.
Basic resistance straw procedure class.

The function that looks up info about a pallet number currently
lives one directory up. I think resistance is the only code that
refers to it. If I can confirm this, I'll move it into this folder.
But ultimately we'd like to try and rely on the DB for this sort
of lookup. This can eventually go in txtprocessor.

Anyways, we need to change the return signature of this function
(which is, by the way, removeStraw::getPallet()). We need it to return
a CPALID in addition to some other info. So do that.

In the resistance GUI itself, beginning some reorganization and
cleaning up, but more importantly: now makes a procedure + session.
Next, should be quick, add worker logins. Then, longer: save
and load resistance measurements.

Just occurring to me that maybe I should add load functionality
to prep?