Save Straw Resistance to DB (#123)
Now save data from the resistance gui to the database.

Follow a similar model as with prep. No need to make straws, straw location, or straw present entries, this time, though -- just resistance measurements in the measurement_ohms table.

Lots of todos: frontend could use a timer and a display of the pallet number and ID. In the backend, should save the measurement method, and a meaningful pass-fail value. Within the GUI, the flow control is wanting -- collect data does too much. Data bound for the database should be validated before committing.

* Better to catch a specific exception.

* Fix a couple of outdated imports.

Discovered that these were problematic as I tried to set up
a test sandbox, feeding the gui fake data.

* As with Prep, remove the not cool main function.

processEvents gets called repeatedly, automatically, when you do
app.exec_().

Logging in and (un)locking the gui happen with emitting pyqt signals.

* Start saving to DB.

Basic resistance straw procedure class.

The function that looks up info about a pallet number currently
lives one directory up. I think resistance is the only code that
refers to it. If I can confirm this, I'll move it into this folder.
But ultimately we'd like to try and rely on the DB for this sort
of lookup. This can eventually go in txtprocessor.

Anyways, we need to change the return signature of this function
(which is, by the way, removeStraw::getPallet()). We need it to return
a CPALID in addition to some other info. So do that.

In the resistance GUI itself, beginning some reorganization and
cleaning up, but more importantly: now makes a procedure + session.
Next, should be quick, add worker logins. Then, longer: save
and load resistance measurements.

Just occurring to me that maybe I should add load functionality
to prep?

* Save worker logins/outs to DB.

Logout button actually logs users out now. Fixed for prep as well.

* Move resistance measurement into its procedure.

It may not be here forever, but for now it's ok.

* Working on DB saving of resistance measurements.

Work in progress.

* Erroneously added a bug line to pallet file.

* Now saves measurements to db. Add timing.

Just make the DB entry objects in the config data function directly.

Not much more critical stuff I can think to do here except add
back in the real data reading from the meters. And then lab test, and
plenty of non-critical to-dos.

One note: doesn't behave particularily well when you re-do/resume
a cpal you've already worked on.

* Big, final reorganization. Clean up. Update comments, header.

* Re-add arduino connection and data-taking stuff.

* Remove erroneous print of pallet filename found.

Though now that I think about it, this should probably be logged.