Leak Refit Program Remakes The Plot (#136)
This is now in addition to spitting out the recalculated leak rate + err.

Tag added is "_refit" instead of "_fit".

Additionally, did more refactoring.

* Break out leak evaluation, plotting into utilities.

Tested for the refit program. Need to make sure still works with
LeakGUI.

* Don't need to save lots of variables by-chamber.

The maps were confusing and unused. Just keep variables in limited loop scope.

* This nearly concludes what has turned out to be a refactor of handleStart.

Here, break out the reading of arduinos and the writing of the data to files
into their own functions.

A lot of slimming down, in general.

* Straw Leak bugfix -- before data is taken running duration was undefined.

* Fix several (hopefully) final bugs.

Working so far. Going to wait for this dummy straw to finish testing
before I finish this branch.

* Another bugfix.