Standalone Automerge (#140)
Fixed the if __name__ == __main__ block of merger.py. Now runs a one-shot automerge.

Intention is to add this to the beginning of mergedown, so that you don't have to open pangui to first trigger a mergedown.

* Switch DM db location lookup to common functions.

Add GetNetworkDatabaseLocation.

* Add a very verbose option with be_verbose.

* Re-wrie the __main__ block to perform a single Automerge.

When you call with python -m guis.common.merger now, you will
execute one automerge (aka Merger::mergeAll).

Seems to work OK. Going to add this to the beginning of mergedown.