Remove straw from position before adding.
The straw_location::addStraw function is broken. It's supposed to give
up if you're trying to add a straw to a position that's already filled.
This check is broken and should be fixed.

So anyway, in the LPAL loader, I now manually remove straws from the
position before adding a new one. Previously, I had accounted for the

Also a bugfix: forgot to return the unfilled positions in that function.