LPAL Loader Bugfixes (#159)
Wasn't passing errors/crashes to the logfile. That's fixed.

The comparison between DB and txt file wasn't working, and it overreported mismatches.

You can now better replace straws that you had previously messed up.

* Fix except hook, skip empty lines when reading.

Except hook to send crashes to the log file.

When reading the csv file to see which positions were filled,
empty lines with just ,, (and not ^M) were crashing the program.
Skip those lines.

* Fix DB's getUnfilled function.

This function previously failed to account for when a straw was
removed from a position and then another (or the same) straw was
re-added to the same position.

New function called getFilled and the diffs it with all positions.

* Remove straw from position before adding.

The straw_location::addStraw function is broken. It's supposed to give
up if you're trying to add a straw to a position that's already filled.
This check is broken and should be fixed.

So anyway, in the LPAL loader, I now manually remove straws from the
position before adding a new one. Previously, I had accounted for the

Also a bugfix: forgot to return the unfilled positions in that function.