Nonsequential steps for processes 6 and 8. (#162)
* Added Preliminary Nonsequential Funcitonality

This kinda sorta works, but it is buggy and currently glitches out on the last step each time. So I shall modify it heavily.

* Progress on loading

Some "scribbles" regarding the setting up of saved checkboxes upon opening pangui for a 2nd time and on.

* Rectified Save Issue

Fixed unknown problem causing issues saving step completion.

* Implemented Save Functionality

Works with the exception of duplicating the saving of first step an obscene number of times and also failing to save the last step. Should be easy fixes, also need to remove debugging comments and clean up the code. Largely functional.

* Added Preliminary Nonsequential Funcitonality

This kinda sorta works, but it is buggy and currently glitches out on the last step each time. So I shall modify it heavily.

* Progress on loading

Some "scribbles" regarding the setting up of saved checkboxes upon opening pangui for a 2nd time and on.

* Rectified Save Issue

Fixed unknown problem causing issues saving step completion.

* Implemented Save Functionality

Works with the exception of duplicating the saving of first step an obscene number of times and also failing to save the last step. Should be easy fixes, also need to remove debugging comments and clean up the code. Largely functional.

* fixed save bug

Caught logical error keeping last step from saving.

* Full Functionality

Got it working for all current cases and put in proper group lists. Will clean up code a bit.

* Added Comments, removed Obfuscation

Cleaned up the code structure and added comments, also removed debugging print statements.

* Added more comments, cleaned logic

Added more comments, modified some poor logic.

* Fixed bug

Fixed it so that current checkbox will get fixed.

* Caught Skipping Bug

Caught and fixed a bug that caused the code to skip from one subgroup to the next.

* Implemented functionality for back to back lists

Made it so that back to back nonsequential lists will check off properly.

* Fix for subgroup enabler function

Fixed it so that the checkbox right after the enabler function will properly be enabled.

* Replaced !=

Replaced != operators with "is not"

* Added comments, description, elif

I added comments and updated the "checkProgress" function description. Also turned an if into an elif, since the two if statements were very closely related.

* Put code back in removed by rebase

Some code was removed during the rebase, so I put it back in.

* Added call to parsesteps

Added one extra call to parsesteps in case it's the first time opening a panel.

* Updated group lists.

Updated group lists to match current steps.

* Added 2nd group list.

Forgot to make sure that both group lists were updated.

* Added compatibility with updated steps list.

Added compatibility with updated steps list. Also fixed the checking off of the last step in the last group. Also put back some files I accidentally deleted in the last commit.

* Removed Typo

Removed a stray "x" that I put in pangui.

* Readded missing files.

Accidentally deleted some files on branch. For wiretensioner and tensionbox. This restores them.

* Removed Print Statement

Removed Print Statement from steps.py.

* Added interim bit of code for edge case.

The way the code works, it's not compatible with some of the old steps checkoff list. This makes it compatible.

* Re-Added Deleted Folder

Straw Folder also Disappeared in commit. Rectified.

* Restored more files.

Restored missing files from wiretensioner, strawtensioner, and tensionbox.

* Added checkbox function

Created a function in stepsList.py to get the next step that has a checkbox, so that ones without checkboxes aren't dealt with. As is proper.

* Removed Print Statements

Removed two print statements utilized for debugging.

* Removed Unnecessary Edge Case

Removed edge case accounting that was determined to be unnecessary.

* Corrected Spelling

Corrected spelling of step.