New Script to Load All Heat Data into DB (#141)
The script still needs some work – specifically edge case checking and more dry runs to make sure it's doing the right thing for all files. See the analogous leak test script and https://trello.com/c/0N3yYeIp.

Finishing the pull because it's not hurting anyone right now.

But finishing this task up is a todo.

* commit 1

Unfortunate. Cannot find the error preventing it from saving. Ack. Not sure what's causing it, though that might be more due to sleep deprivation at this point than it being a difficult problem to solve.

* load with dictionaries

Ok, here's the document. Still doesn't work due to same error, but it's far prettier.

* Functional Update

* Removed first

* Dictionary Based Autoloader Script

This script loads all good heat csv files into the database and copies problem files into the "problem_files" folder.

* Altered Data Structures, iteration

Changed tuples to lists in heat_files_info dictionary. Made iteration cleaner/easier to read.

* Added logger Functionality

Now puts successful loads and a list of problem files into the logger.

* Increased maximum allowed panel number

Increased it due to the fact that we are in the 190s in panels, soon valid panels would throw an error.

* commit 1

Unfortunate. Cannot find the error preventing it from saving. Ack. Not sure what's causing it, though that might be more due to sleep deprivation at this point than it being a difficult problem to solve.

* Removed first