Protect heater from erroneously hitting setpoint early.
We saw a glitch in the arduino get a 1000 deg temp just 1 minute
into heating. This triggered the hold-at-setpoint mode of the
heating script, and therefore the panel didn't heat at 55
for long enough.

So require that the temp is also less than 100 deg. This
should be narrow-enough to dodge most glitches but catch
real temperatures.