Process 1 LPAL Validation (#173)
Modified function "checkLPALS()" in pangui.py. Now it calls the class method _queryStrawLocation() from straw_location.py in order to check that the given lpals exist. If not, it will clear the pertinent text fields and push a message to the user instructing them to merge down. Otherwise, it allows the code to proceed as it did previously.

* Added lpal check call

Added calls to _queryStrawLocation. Additionally, set the lpal boxes to be cleared upon failure, this ensures that bad data won't be saved.

* Added Conditional

Added a conditional to let a valid entry still be saved in the case that there's one valid entry and one invalid entry.