Correctly handle when CPAL is being newly added to DB.
Consolidate should add the CPAL to DB but until it does, this may
be the first time the DB is seeing this CPAL. It can't be created
if the CPALID still has straws on it. Catch that case to remove
the old straws from the CPALID and add the entry.

Upstream of that: if we have CPAL file troubles, just quit.
Hopefully most of them will go away soon when mergedowns for
the CPAL files are eliminated.