Modernize Silver Epoxy GUI (#179)
The usual straw GUI modernization overhaul to save the data into the DB.

In particular, we're just saving to a `procedure_details_silver` table the straw location (CPAL) and the epoxy timer.

With the latest changes to the straw room, we've got new CPALs flying around all over the place, and they're not always being saved to the DB. So this is the first place that the DB will see a lot of these CPALs.

Added some QA checks on the CPAL number inputted.

Not sure if comments can actually get saved in the DB right now. But they should at least get saved to the text file. Editing the pallet internally doesn't work.

* Modernize silver epoxy gui.

Just about done. Details at this point.

In particular: be more careful in purging straws from this
cpal (which could have been typed in in error).

IDK that's all I can think of RN.

* Add comments.

* Don't panic if a CPAL file isn't found.

Previously we crashed (or maybe even silently failed?) when the
CPAL you're trying to SE doesn't exist. Now just tell the user
about this and tell them to mergedown.

* Correctly handle when CPAL is being newly added to DB.

Consolidate should add the CPAL to DB but until it does, this may
be the first time the DB is seeing this CPAL. It can't be created
if the CPALID still has straws on it. Catch that case to remove
the old straws from the CPALID and add the entry.

Upstream of that: if we have CPAL file troubles, just quit.
Hopefully most of them will go away soon when mergedowns for
the CPAL files are eliminated.