DBV Updates May 2022 (#188)
* Added pro 8 typed in leak rate
- gets leak_rate from procedure_details_pan8 and displays in finalqc tab
- also changed "Unknown" to "Not Found" for not found finalqc data

* Added listWidget for pro8 leakTests
- double-clicking the items in it calls the leak rate plot function
- aforementioned funciton (graphLeak) is not done yet
- modified findPro8 to get finalQC leak test details
- modified displayPro8 to display finalQC leak test details
- removed some redundant code

* small comment/debugging changes
-see title of commit

* Ready to start on graphing
- worked on graphLeak(), added database search part
- all that needs to be done is graphing the data

* Saving progress on graphing
- graphs pressure difference so far

* Graphing for final QC leak tests complete
- functions, but most likely could use improvements on appearance
- identified source of trip status bug, see comment on line 1431
- added comments to graphing part of graphLeak()

* Leak test graph visual improvements
- moved diff pressure to left side of the graph
- added gridlines

* Small visual change
- both the subplots for leak test data had different x axis labels
- now they are the same

* One more update - instructions added
- I realized that nobody would know that double clicking a leak test in the list would graph it
- so now there's instructions at the top of the list widget!

The biggest changes exist in facileDBMain.py, but there are also UI changes to the DBV and additions (leak test details) to the PanelData class in panelData.py. There were no changes outside of the dbviewer folder.

The main changes are the addition of leak rate and leak tests to the final QC page in the DBV. The leak rate will be displayed below where the parts are displayed and it's loaded/displayed in the same way (findPro8() and displayPro8()). The leak test details (tag name and duration) get displayed in a list widget next to the parts. They are loaded in there by the same functions as the leak rate. If you double click on an entry in the leak tests list it will launch a pop-up with the graphs for that test. The function responsible for this is graphLeak(). It handles a lot of data (several thousand tuples w/ 7 indexes each) and might make some of the computers lag quite a bit.