Rewrite FindCPAL Function.
It's now called FindCPALContainingStraw, and it no longer
chooses the CPAL with the greatest CPAL number. Instead,
when multiple cpals are found containing a straw, it picks
the one with the straw in the final line of the pallet file.
And when that doesn't resolve the ambiguity, choose the one
with the latest timestamp.