Reduce verbosity of CPAL search function.
While the GUI is running, the running rows' chambers are
constantly being looped through, and the straw info is
constantly being updated in the pallet files.

The update function searches anew, each time it is called,
for the straw's pallet file. Eventually we would like to
probably cache the pallet files, but until then, we must
at least have the search function running efficiently with
low verbosity.

On this branch, I don't think I've made the search function
slower, and I've moved the verbosity back down to ~ zero.

Going to push this now and see how it goes.