Straw Leak CPAL Search Upgrade (#198)
When one straw is found in multiple pallet files, choose the pallet file with the straw in the final line of the file and/or the one with line with the latest timestamp.

Explanation:
CPAL preparing has gotten more involved recently, esp since we started doing the many-retests experiment.

No longer is the pallet with the highest number necessarily the one where a straw should be located.

Make the search-pallet-files-containing-strawname more sophisticated: between two files containing strawname, choose the one which has the straw in the final line of the file. When multiple files satisfy that criterion, use the one with the latest timestamp.

* Rewrite FindCPAL Function.

It's now called FindCPALContainingStraw, and it no longer
chooses the CPAL with the greatest CPAL number. Instead,
when multiple cpals are found containing a straw, it picks
the one with the straw in the final line of the pallet file.
And when that doesn't resolve the ambiguity, choose the one
with the latest timestamp.

* Bugfixes.

Still problems though:update straw function is
called repeatedly, constantly looking up the cpal file
and spamming the logger.

* Reduce verbosity of CPAL search function.

While the GUI is running, the running rows' chambers are
constantly being looped through, and the straw info is
constantly being updated in the pallet files.

The update function searches anew, each time it is called,
for the straw's pallet file. Eventually we would like to
probably cache the pallet files, but until then, we must
at least have the search function running efficiently with
low verbosity.

On this branch, I don't think I've made the search function
slower, and I've moved the verbosity back down to ~ zero.

Going to push this now and see how it goes.