Straw leak bugfix.
I want to return the value of a dict. In one case, that involves
getting the one and only element of the dict. In the second case,
I get the dict value whose key is the max of all such keys. In
this latter case, the max method doesn't give me the whole dict
entry, but instead just what I want: the corresponding value.