Modernize CO2 GUI (#146)
Now saves to the DB, otherwise little functional change.

Gave it the usual treatment under the hood.

* In the process of doing the basic stuff.

Including guis locking, worker credentials, timers, adding DP.

* Lots of fixes, reorganization.

Currently, working on the interaction with the DP. ATM it's
complaining that the CPALID already has straws on it, so it
can't create the CPAL I'm asking it to make.

* Progress on saving to DB.

Added a pallet class function to remove all straws from a pallet ID.

Uncomment the CO2 class in procedures straw.

* Fix up CO2 procedure class.

Important bugfix in CO2GUI! Add comments to procedure.

* Save epoxy batches.

Next: save elapsed time. Then we're basically done.

* Save process duration.

Move stuff around. TODO: prompt user to continue when (a) procedure
already exists or when (b) CPALID has another pallet's straws on it.

* Overhaul resetGUI fn. Prompt user when something goes sideways.

Prompt user to continue or not when (pallet not in DB and another pallet's
straws are on this ID) and when (co2 procedure for this pallet already exists).

resetGUI now kills the whole gui and re-opens it. So you have to log
back in and re-type the three fields. Sorry, but it was the most
straightforward way I could think of to clear the DP and session
completely. Was running into problems when I cleared it manually.

* Bugfix.

* Re-remove updateBoard.

* Cleanup.

* Lost main window import.

* Fix writing to file.

Also, close the old instance before opening a new one
upon successful completion of a gui session.