Single Step Assurance
This update ensures that only one instance of save step execution is saved. The only modified code is in the dataProcessor file in saveStep. It checks for preexisting steps that match the data that you're trying to enter. And it it's already in the db, it simply doesn't enter it.