Straw pallet generator bugfix.
When a new CPAL is being made if the CPALID still has straws, it
would crash. This is fixed now.

Also, add an automerge at the end of the script.

TODOs: When making/retrieving a CPAL or LPAL check whether the
(number,id) pair exists, instead of just the number.