Pmf detection (#213)
Modified consolidate_straws.py and added find_pmf.py. Find_pmf.py includes a function to acquire a list of pmf straws, a function to determine if a straw is pmf, and a main function that's standalone in a loop for manual use. Consolidate_straws.py now uses find_pmf.py's functions to determine and save the number of pmf straws on a consolidated pallet.