Dead Channel Entry Script Now Takes a CSV file (#205)
Deprecated the interactive version of this.

Using it now to upload 85 confirmed bad channels.

Added a uniqueness constraint to the DB table:

```
UNIQUE (
    position,
    procedure,
    wire
)
ON CONFLICT ABORT
```

* Modifying dead chan entry script to handle a csv file.

Right now, just read the CSV file into a data class,
print out the objects.

* Bugfix. Correct way to call elements of a dict.

* TODO make procedure if doesn't exist.

* Many more checks on results of csv parse.

Add timetamps arg to db entry.

Create a procedure for panels if they don't exist.

* Remove deprecated code.

Not sure why it's here...autoformatter? It's unrelated to this branch.

* Don't submit duplicate entries.

I think this is ready to go.